Move GetGithubRelease() to where it belongs
