+++ pkg/ddevapp/addons_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/ddev/ddev/pkg/ddevapp"
 	"github.com/ddev/ddev/pkg/fileutil"
+	"github.com/ddev/ddev/pkg/github"
 	"github.com/ddev/ddev/pkg/globalconfig"
 	"github.com/ddev/ddev/pkg/testcommon"
 	"github.com/stretchr/testify/assert"
@@ -270,7 +271,7 @@ func TestGetGitHubRelease(t *testing.T) {
 
 	// Test getting latest release
 	t.Run("GetLatestRelease", func(t *testing.T) {
-		tarballURL, version, err := ddevapp.GetGitHubRelease("ddev", "ddev-redis", "")
+		tarballURL, version, err := github.GetGitHubRelease("ddev", "ddev-redis", "")
 		require.NoError(t, err, "Should successfully get latest release")
 		require.NotEmpty(t, tarballURL, "Tarball URL should not be empty")
 		require.NotEmpty(t, version, "Version should not be empty")
@@ -279,7 +280,7 @@ func TestGetGitHubRelease(t *testing.T) {
 
 	// Test getting specific version
 	t.Run("GetSpecificVersion", func(t *testing.T) {
-		tarballURL, version, err := ddevapp.GetGitHubRelease("ddev", "ddev-redis", "v1.0.4")
+		tarballURL, version, err := github.GetGitHubRelease("ddev", "ddev-redis", "v1.0.4")
 		require.NoError(t, err, "Should successfully get specific version")
 		require.Equal(t, "v1.0.4", version, "Should return requested version")
 		require.NotEmpty(t, tarballURL, "Tarball URL should not be empty")
@@ -287,14 +288,14 @@ func TestGetGitHubRelease(t *testing.T) {
 
 	// Test non-existent repository
 	t.Run("NonExistentRepo", func(t *testing.T) {
-		_, _, err := ddevapp.GetGitHubRelease("ddev", "non-existent-repo", "")
+		_, _, err := github.GetGitHubRelease("ddev", "non-existent-repo", "")
 		require.Error(t, err, "Should fail for non-existent repository")
 		require.Contains(t, err.Error(), "unable to get releases", "Error should mention inability to get releases")
 	})
 
 	// Test non-existent version
 	t.Run("NonExistentVersion", func(t *testing.T) {
-		_, _, err := ddevapp.GetGitHubRelease("ddev", "ddev-redis", "v999.999.999")
+		_, _, err := github.GetGitHubRelease("ddev", "ddev-redis", "v999.999.999")
 		require.Error(t, err, "Should fail for non-existent version")
 		require.Contains(t, err.Error(), "no release found", "Error should mention no release found")
 	})
@@ -302,7 +303,7 @@ func TestGetGitHubRelease(t *testing.T) {
 	// Test real addon with dependencies
 	t.Run("RealAddonWithDependencies", func(t *testing.T) {
 		// Test ddev-redis-commander which depends on ddev-redis
-		tarballURL, version, err := ddevapp.GetGitHubRelease("ddev", "ddev-redis-commander", "")
+		tarballURL, version, err := github.GetGitHubRelease("ddev", "ddev-redis-commander", "")
 		require.NoError(t, err, "Should successfully get ddev-redis-commander release")
 		require.NotEmpty(t, tarballURL, "Tarball URL should not be empty")
 		require.NotEmpty(t, version, "Version should not be empty")
